+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added Helm Chart
 - Support yaml as configuration import format. (`--import.file-type=yaml`)
 - In some situations if Keycloak gives 400 HTTP error, pass error message from keycloak to log.
+- Allow updating builtin flows and executions (keycloak allows to change some properties)
+- Remove authentications config from keycloak if not defined in realm
 
 ### Changed